feat(Express): move socket to service and add last data sending
